requirements.txt

fastapi==0.80.0
Keras==2.2.4
numpy==1.18.5
opencv-python==4.4.0.40
pandas==1.1.3
Pillow==9.2.0
tensorflow==2.3.4
uvicorn==0.18.3
python-multipart==0.0.5
